c pracrice

+++ 6.c
@@ -1,7 +1,7 @@
 #include<stdio.h>
 int main()
 {
-    int fact =1 ,i=1,n;
+    long fact =1 ,i=1,n;
     printf("enter number to find factorial");
     scanf("%d",&n);
     while(i<=n)
